Let an external OIDC provider be set with flags
This commit adds flags for an external OIDC provider, and changes how
the middleware is constructed accordingly.

The external OIDC options get their own field. `core/client`
obligingly lets you give a prefix for flags registered by HTTPOptions,
when constructing it, so we can have both options for using the
identity service, and options for using an external provider. The main
one we care about is the host, but the whole lot (CA secret) can be
provided.

In this service, the choice is between:
 - use internal state to authenticate all tokens; or,
 - use internal state for encrypted tokens (assumed to be minted
   here), and the external provider for other tokens.

In other services, it'll be between
 - just the identity service; or
 - the identity service or an external provider

So server.go elsewhere will have a similar but distinct if-else to
construct an authorizer.