Account for external, encrypted access tokens
Up to this point my assumption was that if anyone was doing encrypted
tokens, it was the identity service. But Auth0 uses JWEs with direct
encryption -- i.e., essentially opaque tokens.

Usually that would make it much trickier to determine who to ask about
a given token, but they _also_ include a (non-standard) "iss" field in
the header, which is unencrypted. I don't know if other providers do
this; if they don't, we'll need another tactic.