Remove non-standard UpdatedAt field
The UpdatedAt field in Userinfo is not standard, and it turns out some
providers send a string, and at least one provider (Zitadel) sends a
number. This means it fails to deserialise, when we call /userinfo.

We don't appear to use it (in UNI anyway), so rather than figuring out
a scheme for deserialising string-or-number, just remove it for now.