Add "hybrid" authentication
This commit adds an implementation of Authorizer that dispatches to
either a local authenticator, or a remote authoenticator, depending on
what it can detect about the access token.

The motivation for this is to be able to handle tokens issued by the
identity service as well as tokens issued by an external platform --
so long as they can be distinguished.