Make token detection more precise
We have to detect whether a token is most likely from the remote
issuer, or from our local issuer. However: both the UNI identity
service and at least one external provider (Auth0) use JWEs for access
tokens, so the fact of a token being a JWE is not enough to detect
where it should go.

Luckily, Auth0 at least includes a (non-standard) `"iss"` field in the
header of its encrypted tokens. This means we can cover at least that
possibility, by checking for `"iss"` and dispatching based on whether
it matches the given local or external issuer.

If `"iss"` is absent in a JWE, assume it's local (UNI JWEs don't have
`"iss"` in the header). This is the weakest part of the detection,
since it's likely other providers could use JWEs, but include an
issuer in the header.

Yet other providers will send opaque tokens: assuming those are remote
covers off another possibility.