Include account type in userinfo response
I'm trying to decouple authorisation from identity, so that (at least
in principle) an authorisation service can be run separately to the
identity service. One bit to detangle is that `RBAC.GetACL()` relies
on knowing the kind of account represented by an access token.

This works at present because the handler that calls GetACL is behind
middleware (pkg/middleware/openapi/local) which puts that information
in the context. But since this relies on running in the same place as
the token issuer, which has decryption keys, a decoupled authorisation
service would not be able to see that information.

So: include the account type amongst the claims returned by /userinfo,
so that it will be visible to the middleware, whether called remotely
or in-process with the issuer.

This field will _not_ be present in /userinfo returned by other
providers; but that can be worked around:

 - assume a /userinfo response without the field is a user (works if
   system and service accounts are only internal to UNI);

 - add the claim to /userinfo results from the provider, by using a
   platform-specific extension mechanism (platforms like Auth0 let you
   add claims into the ID tokens)