Use the new acctype claim in middleware
This standardises how system and service accounts are signified: the
claim is in the /userinfo result, as calculated by the issuer, and
either returned directly or via the /userinfo endpoint. Now, both the
local and remote authorizer check what particular token represents by
looking at the userinfo claims.

In pkg/middleware/openapi/local we are running locally with the
internal OIDC implementation, and there are some additional checks to
do (like: did the system user present a valid certificate along with
the access token?) -- I have left those in place.