External Finalizers (#294)
At present we rely on resources living in an identity managed project
namespace in order to get cascading, blocking deletion semantics.  This
is bad for a number of reasons: a) the service needs to live in the same
cluster, b) the service needs to do a bunch of indirection to find where
the resource exists, and has knowledge of identity's layout.

Looking forward, what we want is to separate concerns, so identity can
do what it wants, and the external service can do what it wants with
very little coupling.  Taking compute as an example, when we create a
cluster, we want to mark the project as "in use" by the cluster,
blocking deletion until we are done with it.  Cluster deletion can
either be explicit, or implicit via a message bus (e.g. Kubernetes
watch, or a message queue).  In either case, once cleanup has completed
we can remove the reference from the project in order to allow deletion.

We use an API, rather than direct project access to enforce data
boundaries and reinforce the whole decoupling drive.

As a side effect, for compute in particular, all clusters can be moved
to a single namespace, meaning roles can be tightened.  Resource lookups
can be done directly by ID, rather than walking a "directory" structure
first.  In whole this leads to increased performance, security, and
scalability.