Test remote authorizer
This adds a test for the remote authentication implementation. There
is a lot of setup -- mainly creating certificates -- but the payoff is
that we can verify that it ends up calling the right endpoint in the
right way.

The server responding to /userinfo and the discovery endpoint does the
same as the "real" identity server; it could _be_ the real identity
server, except that drags in several other dependencies that would
have to be constructed. This fake one is simple enough.