Add organisation IDs to userinfo response
One obstacle in trying to disentangle authorisation from the IdP part
of UNI is that the authorisation server needs to know organisation
memberships, in order for it to calculate ACLs. If the memberships
(along with the users) are recorded in another IdP, it can't access
them.

There's different ways to solve this:

- have a parallel implementation of authorisation that works with the
  IdP of choice; or,
- integrate IdP management APIs into the authorisation service, to be
  switched on with flags;

but the cleanest one is to include the membership information in the
ID token (or rather, the /userinfo response) as a custom claim. Any
external IdP service is likely to have mechanisms for putting custom
claims into tokens.