Migrate UNI RBAC groups to use Subjects field
In a previous commit, the Group type was changed so that it uses the
user subject as the key for refering to users, rather than the ID of
the OrganizationUser object representing membership of the
organisation. In running systems, the UserIDs field will be populated,
so we need to port those to use subjects instead.

This does the simplest version of that, by running a func from main.go
on startup. The func will update the subjects field. It only makes
sense to do this once, so it will only port objects if the subject
field has no entries. The subject is obtained by looking at the
OrganizationUser object referenced in the organisation namespace; then
finding the corresponding User record in the "global" namespace
(usually `unikorn-identity`).