Migrate UNI RBAC groups to use Subjects field
In a previous commit, the Group type was changed so that it uses the
user subject as the key for refering to users, rather than the ID of
the OrganizationUser object representing membership of the
organisation. In running systems, the UserIDs field will be populated,
so we need to port those to use subjects instead.

This migration is intended to be run once manually ahead of deployment.
It only makes sense to run it once, so it will only port objects if the
subjects field has no entries. The UserIDs field is left alone so that
processes still using it do not break.

The subject is obtained by looking at the OrganizationUser object
referenced in the organisation namespace; then finding the
corresponding User record in the "global" namespace (usually
`unikorn-identity`). The email is used as both the subject record's
email value (for display), and for the ID -- usually you'd expect the
ID to be something opaque, but since UNI issues tokens with the email
as the subject, we use that as the ID.