Use the user API handler to create test objects
Since I'm trying to make a change in a backward-compatible way, I am
going to use the API handler rather than the Kubernetes client for
creating users. If the change goes well, the tests should pass without
alteration -- because "backwarbcompatible.

This ends up saving lines, because the handler does all the object
creation.