Add Group.Spec.Subjects field
This commit adds a field for referring to users directly. If using an
external IdP, we will not have the User and OrganizationUser records
to refer to (in .UserIDs).

This is to some extent backward-compatible: if the callers keep using
the UserIDs field, then the server will keep populating it -- while also
populating the Subjects field, to use in calculations.

The part that isn't backward-compatible is that once you send a
request with Subjects supplied, then UserIDs is forgotten about. These
fields are not interchangeable, because the subjects can refer to
users somewhere else.