Add Networking/Storage Quotas (#315)
Networks, especially for bare metal instances, will require a phyisical
network segment (VLAN/VXLAN/etc) to provide L2 isolation.  For the
former in particular, that's limited to a mere ~4096, which can very
easily lead to exhaustion and denial of service attacks if not
constrained.  This must be in place before we start on user-provisioned
networks in the region service.  Additionally adds in a public IPs quota
to limit starvation of this most scarce of resources.  And also storage.

Note that the clusters quota is ostensibly useless now as that was used
as a very rough way to prevent VLAN starvation.