Move user database operations to their own abstraction
This commit separates the user- and membership-oriented operations
from the RBAC queries. The eventual aim is that RBAC does not depend
on running with the user database, since that precludes using an
external provider for authentication.

Mostly this is a simple case of moving methods to a new abstraction in
pkg/userdb, then fixing up the callers.

One wrinkle is that the Onboard method in oauth2 needs to produce a
"super user" context, and that can only be done with understanding of
RBAC; but at least this dependency runs from oauth2 to RBAC, and not
from RBAC to userdb.