Rewrite authorisation without user database
This commit removes the user database methods from RBAC, which have
been moved to pkg/userdb, and rewrites the GetACLs method to use the
information in the userinfo instead. This means the RBAC no longer
relies on the service account, user and org'user records; so it will
work with an external provider, so long as that provider can be
persuaded to include the custom claims in userinfo.

This also needs the Subject field to be populated in Group
objects. For a live system, this would mean running some migration to
populate those fields from the OrganizationUser records.