Test remote authoriser (#310)
* Test remote authorizer

This adds a test for the remote authentication implementation. There
is a lot of setup -- mainly creating certificates -- but the payoff is
that we can verify that it ends up calling the right endpoint in the
right way.

The server responding to /userinfo and the discovery endpoint does the
same as the "real" identity server; it could _be_ the real identity
server, except that drags in several other dependencies that would
have to be constructed. This fake one is simple enough.

* Factor out MTLSServer abstraction

It's likely we will want to do this CA-plus-server-plus-client
certificates setup in a few places, so factor it out into its own
abstraction, and make it look like httptest.Server.

* Rewrite remote auth test to use MTLSServer

* Boil down test boilerplate

 - Factor out constructing the Auth'Info struct
 - make invalid tokens a table test
 - remove a test that didn't test anything new
 - test caching by checking that only one upstream request is made

* Add authorisation test for plausible, but bad token

That is: a token that is definitely a signed JWT, just not one we
trust.

This adds a bunch of lines for not a huge benefit, IMO, so we could
take or leave it.

* Move helpers to end of file