Add Quota Allocation Client (#319)
This code is duplicated across compute, Kubernetes and now region
services, which leads to a maintenance nightmare, and things not
aligning causing bugs.  Move this to a central, shared location to fix
once, fix everywhere, and also allow us to modify the APIs in one
place, which is going to be important for serialization of requests to
enforce strong consistency and avoid race conditions.