Remove userdb from organization handlers
The organization list endpoint does, in effect, an inline
authorisation check: you only see the organisations that you are a
member of, unless you have the global permission
`identity:organizations`.

To get the same effect without relying on user membership, as with
decoupling RBAC proper we can check the organization IDs given in the
userinfo. The basic case is: show organisations for which the ID is
mentioned in the userinfo.

There's some special cases:

 - if you pass your own email, it'll give you just the orgs to which
   you are a member; this might matter if you are a platform-admin and
   can otherwise see all orgs

 - if another email is passed, then it's supposed to get the organisations
   for that user, given adequate permissions. But we don't have that
   information! So this behaviour **is not preserved**