Expand group subjects to include provenance
This changes the type of a group's subjects to be `{ID, Email,
Issuer}`. The motivation is to give the system, user interfaces, and
troubleshooting humans, some idea of provenance.

The key for the records is then {ID, Issuer}, with the email
effectively being a display name. For this reason, when adding and
removing subjects from groups, both the ID and Issuer are
considered. (It's not _that_ likely that subject IDs will collide --
only really if two user databases use the email as the subject, which
UNI does, but no-one else.)

The UNI user database is a special case, in that it uses the User
email as the subject ID. The email is used as the subject because UNI
issues tokens with the email as the subject; when RBAC is decoupled
entirely from the user database, it will need to be able to look up
group subjects by whatever is in the token.