Standardise how Issuer URL is supplied
In many places, an issuer URL is needed for minting tokens. Most of
these are in response to HTTP requests, and use the request Host
header as a shortcut: the rationale would be, if the request was
routed here, this must be the issuer.

However, the server has a flag for setting the URL to use in the
well-known endpoint -- and any token that is expected to validate at
the server should use the same URL. To enforce so, this commit passes
the issuer URL value through to everything dealing with token Issuer
claims, so those components have the same value whatever an individual
request suggests. (To be clear: I don't think there's a danger of this
being a security hole, so much as it's a trap leaving it to
coincidence.)

That includes the Group client, which needs to put an issuer URL in
the Subject records it synthesises.