Decouple RBAC Group definition from user database (User and OrganizationUser) (#314)
* Add test for ACL invariants

This adds a test that ACLs are calculated correctly, given a
particular set up of group membership.

The aim of this test is that, with the obvious necessary adjustments
to the setup, it should still pass after migrations, demonstrating
that the migration has not affected ACL calculation.

* Use the user API handler to create test objects

Since I'm trying to make a change in a backward-compatible way, I am
going to use the API handler rather than the Kubernetes client for
creating users. If the change goes well, the tests should pass without
alteration -- because "backwarbcompatible.

This ends up saving lines, because the handler does all the object
creation.

* Add Group.Spec.Subjects field

This commit adds a field for referring to users directly. If using an
external IdP, we will not have the User and OrganizationUser records
to refer to (in .UserIDs).

This is to some extent backward-compatible: if the callers keep using
the UserIDs field, then the server will keep populating it -- while also
populating the Subjects field, to use in calculations.

The part that isn't backward-compatible is that once you send a
request with Subjects supplied, then UserIDs is forgotten about. These
fields are not interchangeable, because the subjects can refer to
users somewhere else.

* Maintain subjects field when updating users

This modifies another place that groups are updated: when user records
are created, updated or deleted. Since this only happens when the UNI
user database is in use, both .UserIDs (refering to OrganizationUser
objects) and .Subjects are updated.

* Make lint happy (cosmetic)

Superficial changes so that the linter passes this.

* Expand group subjects to include provenance

This changes the type of a group's subjects to be `{ID, Email,
Issuer}`. The motivation is to give the system, user interfaces, and
troubleshooting humans, some idea of provenance.

The key for the records is then {ID, Issuer}, with the email
effectively being a display name. For this reason, when adding and
removing subjects from groups, both the ID and Issuer are
considered. (It's not _that_ likely that subject IDs will collide --
only really if two user databases use the email as the subject, which
UNI does, but no-one else.)

The UNI user database is a special case, in that it uses the User
email as the subject ID. The email is used as the subject because UNI
issues tokens with the email as the subject; when RBAC is decoupled
entirely from the user database, it will need to be able to look up
group subjects by whatever is in the token.

* Standardise how Issuer URL is supplied

In many places, an issuer URL is needed for minting tokens. Most of
these are in response to HTTP requests, and use the request Host
header as a shortcut: the rationale would be, if the request was
routed here, this must be the issuer.

However, the server has a flag for setting the URL to use in the
well-known endpoint -- and any token that is expected to validate at
the server should use the same URL. To enforce so, this commit passes
the issuer URL value through to everything dealing with token Issuer
claims, so those components have the same value whatever an individual
request suggests. (To be clear: I don't think there's a danger of this
being a security hole, so much as it's a trap leaving it to
coincidence.)

That includes the Group client, which needs to put an issuer URL in
the Subject records it synthesises.