Support Unscoped ACLs
Newer APIs recognize that organization and project IDs are unnecessary
in the path most of the time, especially when accessing a resource by
ID, and then you can derive those from the resource itself after reading
it.  In other situations you can derive it from a linked resource e.g. a
server is attached to a network from which you can inherit the IDs.  The
only time they are really needed to be provided is for a top level
resource, on creation, and those can be passed in the body as immutable
paramters ostensibly.

This leaves us in a world where the usual ACL path (which requires an
organization ID) doesn't work for non-admin users.  The fix however is
quite simple, rather than lookup a known organization ID, look for all
possible IDs that the user is a member of and return an ACL containing
all of those and any nested project permissions.

This lives along side the existing RBAC method allowing a seamless and
backwards compatible upgrade path.  Internal RBAC checks can be migrated
straight away.