Support both UserIDs and Subjects in group API (#326)
Populate (org) userIDs when subjects are supplied

Previously (#314), the groups handler accepted both UserIDs or
Subjects, but if supplied Subjects it would not bother filling out
UserIDs. This complicates porting external callers, because they all
have to move at once.

This commit adds code to do the complement: fill in UserIDs when
supplied Subjects. It can only do that when the subjects have the
internal issuer (meaning the Subject points at a user in the internal
database). But this is enough to port the API callers over safely.