Add Lister Helpers (#328)
When we list resources, especially with the new APIs that don't require
an organization and project in the path, we would have to list literally
every resource in the system, then filter out using ReBAC, which is
horribly inefficient and, more pertinently, pretty insecure.  Instead we
should use label selection and the joys of database indexes to do the
heavy lifting.

We do provide query selectors for organization and project to limit the
scope, but there is no requirement to actually use these.  Now we have
unscoped ACLs however, the ACL itself contains all the organizations and
projects the user has access to, so we can do a number of things.  For a
start we can mandate a default organization selection based on the ACL,
so you're never getting anything that belongs to someone else.
Secondly, when you do specify a selector, we can constrain this to only
organizations in the ACL, thus preventing attacks where you input the
organization ID of another user.

The same applies to projects.