Serialise allocation decisions
When requests for allocations come in concurrently (or near enough),
they can each conclude their allocation is valid, even if in total
they exceed a quota. This results in the allocation vs. quota reaching
an inconsistent state.

The simplest way to avoid making mistakes due to concurrent allocation
decisions is to serialise them. And the simplest way to serialise the
decisions is to have a mutex around the decision making. Here, I have
done the minimally intrusive version of this, which is to construct a
sync.Mutex in a central place (the Handler) and supply it to the code
that makes allocation decisions (the allocations.Client).

This would be sufficient, except that the handler uses a cached
(Kubernetes API) client. This means that another client can see a
stale, cached state and make another inconsistent decision. The other
part of solving this, then, is to use an uncached ("direct") client
for write operations, so that they make decisions based on the live
state.

It's a little awkward constructing the direct client and supplying it,
but at least it's a fairly obvious kludge.