Test concurrent allocations
These tests demonstrate that the critical section (using the mutex)
prevents the allocations from reaching an inconsistent state.

Because of the way the client is now constructed, it is not easy to
run this test against previous code; but it's possible to comment out
the mutex.Lock() and see the tests fail.