Separate OAuth2 Errors (#343)
THis makes a separate oauth2 error type specifically for oauth2 only.
This means that the error type defined in core is for standard APIs only
and can change independently of our OIDC implementation.  This is
basically a like-for-like copy.  The one change is that the server_error
type is now a default catch all if an explcit error was not returned by
the oauth2 handler.