Use self-contained schema for openapi middleware tests
This removes the dependence on the central schema (less chance of an
inadvertant break); and, it means we can put things in the test schema
that don't appear in the real schema -- as I want to do with a tag for
skipping request body validation.

Less dramatically, use httptest.ResponseRecorder instead of a
homebaked ResponseWriter stand-in.