Let operations skip request body validation
On rare occasions, we want to skip request body validation. An example
is servicing image upload requests, for which the request body can be
mega- or gigabytes long. The validation lib (kin-openapi) will read
the whole request body -- both of a requestBody schema, and also for
security validation.

So, give operations (e.g., POST /upload) a way to opt out of request
body validation. This could reasonably be done in a few ways:

 - supply a list of opted-out paths to the middleware;
 - look at the content type

... but I felt the best way was in the schema itself, so I've used a
special tag.

There's an option for skipping request body validation, but it'll be
read anyway for security validation. Therefore `request.Body` is set
to `nil` during validation -- it's the only way to be sure.