Use Crypto a Bit Less...
We were being overly cautious with the use of crypto when propagating
user principal information.  We only trust the input when a client
certificate is presented and validated any way, so the use of RSA PSS
was a little over kill.  So much so that the salted signature generation
actually caused clients to exhaust the cryptographically secure entropy
and totally kill performance!