Automatic Token Rotation and Client Caching (#346)
The current model, for every request, will get a token, then create a
new client.  The first problem is unnecessarily getting a token on every
call (a problem that's solved with SPIFFE...), so what we do is attach
the issuer to the client, this allows it to cache the token and trigger
a refresh on (or before) expiration.  This then means we can have a
single instance of the client in each handler which facilitates the
reuse of the underlying HTTP client's connection pool.  This then means
through the magic of keep alives that we don't need to the TLS handshake
every call.

We apply a similar connection pool caching strategy for use with the
OIDC userinfo and ACL calls, which improves authn/z performance.

The clients themselves receive a bump in the number of connections that
are allowed which gives another relatively sizable boost.

Now the test here was doing 10 concurrent image lists via the compute
service.  That started off with each request taking ~10s to complete.
it's now down to 1-2s at worst.  I'm beginning to suspect the ingress
controller is stalling the requests between when we make the call, and
when it's actually deivered to the remote, but not found an answer quite
yet.