Remove Response Validation (#351)
Responses can be large, and adding in the HTTP snoop stuff is quite
expensive.  We make developers stomach the pain of experiencing a panic
if their stuff doesn't validate, so it doesn't fly under the radar.
Provided they are testing in the first instance... In production mode we
just trim the fat entirely, yielding 200% throughput.