Add types and routes for organisation-scope allocations
To date, allocations have been tied to infra resources, and these are
scoped to projects. Now, we need to be able to request allocations
that are scoped to an organization, rather than a project, for
resources that are organisation-wide, like VM images.

I've taken this opportunity to rename the array type using `List`, so
the schema and the generated types are not so easily mistaken for the
singular type.

When asking for a list of allocations, it's done at the level of an
organisation. But now some allocations will belong to org-level
resources, rather than project-level resources, and we have to somehow
mix those two in the list. For that purpose, the new schema
`allocationResourceMetadata` is added; it has an optional projectID
field, so it can be either project or org scoped.

(I have used the oapi-codegen flag "old-merge-schemas" 1. to work
around a bug
(https://github.com/oapi-codegen/oapi-codegen/issues/632); and 2. so
the core type gets embedded, which makes constructing the values a lot
easier than it would be).

Co-Authored-By: alansy-nscale <alan.sy@nscale.com>